backend/pyproject.toml
[project]
name = "justicetranscribe_backend"
version = "0.1.0"
description = "FastAPI backend for justicetranscribe"
requires-python = ">=3.9"
dependencies = [
     "fastapi[standard]>=0.112.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "sqlmodel>=0.0.16",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
]

